Refactor tool_shed.galaxy_install.tool_dependencies into galaxy.
TODO:
