+++ lib/galaxy/tool_shed/galaxy_install/tool_dependencies/env_manager.py
