+++ lib/galaxy/tool_shed/galaxy_install/tool_dependencies/recipe/env_file_builder.py
