+++ lib/galaxy/tool_shed/galaxy_install/tool_dependencies/recipe/recipe_manager.py
@@ -1,7 +1,7 @@
 import logging
 
-from tool_shed.galaxy_install.tool_dependencies.recipe import step_handler
-from tool_shed.galaxy_install.tool_dependencies.recipe import tag_handler
+from galaxy.tool_shed.galaxy_install.tool_dependencies.recipe import step_handler
+from galaxy.tool_shed.galaxy_install.tool_dependencies.recipe import tag_handler
 
 log = logging.getLogger(__name__)
