+++ lib/tool_shed/galaxy_install/install_manager.py
@@ -10,14 +10,14 @@ from six import string_types
 from sqlalchemy import or_
 
 from galaxy import exceptions, util
+from galaxy.tool_shed.galaxy_install.tool_dependencies.recipe.env_file_builder import EnvFileBuilder
+from galaxy.tool_shed.galaxy_install.tool_dependencies.recipe.install_environment import InstallEnvironment
+from galaxy.tool_shed.galaxy_install.tool_dependencies.recipe.recipe_manager import StepManager
+from galaxy.tool_shed.galaxy_install.tool_dependencies.recipe.recipe_manager import TagManager
 from galaxy.tool_util.deps import views
 from tool_shed.galaxy_install.datatypes import custom_datatype_manager
 from tool_shed.galaxy_install.metadata.installed_repository_metadata_manager import InstalledRepositoryMetadataManager
 from tool_shed.galaxy_install.repository_dependencies import repository_dependency_manager
-from tool_shed.galaxy_install.tool_dependencies.recipe.env_file_builder import EnvFileBuilder
-from tool_shed.galaxy_install.tool_dependencies.recipe.install_environment import InstallEnvironment
-from tool_shed.galaxy_install.tool_dependencies.recipe.recipe_manager import StepManager
-from tool_shed.galaxy_install.tool_dependencies.recipe.recipe_manager import TagManager
 from tool_shed.galaxy_install.tools import data_manager, tool_panel_manager
 from tool_shed.tools.data_table_manager import ShedToolDataTableManager
 from tool_shed.util import basic_util, common_util, encoding_util, hg_util, repository_util
