+++ packages/app/setup.py
@@ -64,16 +64,16 @@ PACKAGES = [
     'galaxy.tools.util',
     'galaxy.tools.util.galaxyops',
     'galaxy.tool_shed',
+    'galaxy.tool_shed.galaxy_install',
     # TODO:
-    # 'galaxy.tool_shed.galaxy_install',
     # 'galaxy.tool_shed.galaxy_install.datatypes',
     # 'galaxy.tool_shed.galaxy_install.grids',
     # 'galaxy.tool_shed.galaxy_install.metadata',
     # 'galaxy.tool_shed.galaxy_install.migrate',
     # 'galaxy.tool_shed.galaxy_install.migrate.versions',
     # 'galaxy.tool_shed.galaxy_install.repository_dependencies',
-    # 'galaxy.tool_shed.galaxy_install.tool_dependencies',
-    # 'galaxy.tool_shed.galaxy_install.tool_dependencies.recipe',
+    'galaxy.tool_shed.galaxy_install.tool_dependencies',
+    'galaxy.tool_shed.galaxy_install.tool_dependencies.recipe',
     # 'galaxy.tool_shed.galaxy_install.tools',
     # 'galaxy.tool_shed.galaxy_install.utility_containers',
     'galaxy.tool_shed.util',
